Structural chagnes
Massive refatoring to model - controller design pattern, now routes
only redirect to the controllers which operate function from
models section which use utility

Sensitive data was packed in the .env
