--- config.js
@@ -1,4 +1,0 @@
-// Later set ENV instead of this file
-module.exports = {
-  secretKey: process.env.JWT_SECRET || "dafj234532jlkdsADFIHAoh35;DJF",
-};
