--- pass.js
@@ -1,3 +1,0 @@
-const password = "FgsGt5689Hr";
-
-module.exports = password;
